Fixed tests. Working on Linux.
